Add widgetType in Tag-Af mapping.
